sixth commit
the main part of this commit would be  editing some user profile and upload the real video and connect the video and user in the database so that we can know which video is belongs to which user.
second part is done with webpack which is a skeleton part before doing scss.